Add files via upload
✨ New Achievement Unlocked! ✨

I just created a simple student card component using React Props. 🎓 The card displays the student's name, roll number, and the day.

Props make it easy to pass data between components in React, and this project was a great hands-on way to understand their power! 🚀

Check out the video and code for more details. 💻👇

🔗 [Your Video Link] 🔗 [Your Code Link]

Thank you to all my mentors for their guidance and support! 🙏

#React #Props #FrontendDevelopment #LearningJourney #NextStep #ReactComponents #WebDevelopment